requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pillow==10.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.8.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.2
requests
urllib3
schedule==1.2.2
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.40
starlette==0.44.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.13.2
urllib3==2.5.0
uvicorn==0.33.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
markdownify
reportlab
apscheduler 
psutil
chardet
python-docx
pytz